+++ Model/ImportExport/Processor/Export/Node/Tree.php
@@ -24,6 +24,7 @@ class Tree
     {
         $tree = [];
         $childNodesClusters = [];
+        $pendingNodesChildClusters = [];
 
         foreach ($nodes as $node) {
             $nodeId = $node->getId();
@@ -34,10 +35,19 @@ class Tree
 
             if (!$parentId) {
                 $tree[$nodeId] = $nodeData;
+
+                if (isset($pendingNodesChildClusters[$nodeId])) {
+                    $tree[$nodeId][ExtendedFields::NODES] = $pendingNodesChildClusters[$nodeId];
+                }
+
                 continue;
             }
 
             $childNodesClusters[$nodeId] = $nodeData;
+
+            if (isset($pendingNodesChildClusters[$nodeId])) {
+                $childNodesClusters[$nodeId][ExtendedFields::NODES] = $pendingNodesChildClusters[$nodeId];
+            }
 
             if (isset($tree[$parentId])) {
                 $tree[$parentId][ExtendedFields::NODES][$nodeId] = &$childNodesClusters[$nodeId];
@@ -48,6 +58,8 @@ class Tree
                 $childNodesClusters[$parentId][ExtendedFields::NODES][$nodeId] = &$childNodesClusters[$nodeId];
                 continue;
             }
+
+            $pendingNodesChildClusters[$parentId][$nodeId] = &$childNodesClusters[$nodeId];
         }
 
         return $this->reindexTreeNodes($tree);